requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0.post1
python-dotenv>=1.0.1
python-multipart>=0.0.7
keyboard>=0.13.5
# pyaudio>=0.2.14
pyOpenSSL>=24.1.0
